+++ dictionary/client.js
@@ -4,3 +4,10 @@
 Template.registerHelper('dict', function(name, defaultValue) {
 	return orion.dictionary.get(name, defaultValue);
 });
+
+/**
+ * Is the dictionary active
+ */
+Template.registerHelper('isDictionaryActive', function() {
+	return orion.dictionary.isActive();
+});