+++ users/client.js
@@ -2,5 +2,5 @@
  * Check if the logged in users has permission for something.
  */
 Template.registerHelper('doIHavePermission', function(key) {
-	return Meteor.user().hasPermission(key);
+	return !!Meteor.user() && Meteor.user().hasPermission(key);
 });